Improvements to the build system and the resulting shared object file
The resulting .so file is now loadable by the Qt core, and our implementation of
createDevice() is properly called. The failure in loading was caused by the .so
having dependencies to libboost-system which weren't listed in the .so.

Also introduce a number of changes :
 * Import cannelloni directly from the github repository and use the files we
   need instead of relying on the user to have libcannelloni-common in her
   system.
 * Add Qt5 and Qt5SerialBus to the default include path in order not to type the
   library names in include directives.
 * Throw out QtCannelloniCanBusPlugin.h and make a self-sustainable cpp file out
   of it.
 * Require using double-quotes when including headers/files in the local
   directory.
