+++ CMakeLists.txt
@@ -1,21 +1,22 @@
-cmake_minimum_required(VERSION 2.6)
-project(qtCannelloniCanBusPlugin)
+cmake_minimum_required(VERSION 3.1)
+project(qtCannelloniCanBusPlugin CXX)
 
-ADD_DEFINITIONS(
-  -std=c++14
-  -Wall
-  )
-
-set(CMAKE_INCLUDE_CURRENT_DIR ON)
+set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_EXTENSIONS OFF)
+#set(CMAKE_INCLUDE_CURRENT_DIR ON)
 set(CMAKE_AUTOMOC ON)
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall")
 
-find_package(Qt5Core)
+find_package(Qt5Core REQUIRED)
+find_package(Qt5SerialBus REQUIRED)
+find_package(Boost REQUIRED system)
 
-include_directories(../)
+include_directories(${Boost_INCLUDE_DIRS} ${Qt5Core_INCLUDE_DIRS} ${Qt5SerialBus_INCLUDE_DIRS} ${CMAKE_BINARY_DIR})
 
 add_library(${PROJECT_NAME} SHARED
   QtCannelloniCanBusPlugin.cpp
   CannelloniCanBackend.cpp
+  cannelloni/parser.cpp
   )
 
-target_link_libraries(${PROJECT_NAME} Qt5::Core)
+target_link_libraries(${PROJECT_NAME} Qt5::Core Qt5::SerialBus ${Boost_LIBRARIES})
